workflow-notifier/requirements.txt
slack-sdk==3.14.1
click==8.0.3
strictyaml==1.6.1
